% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crimson-text.R
\docType{data}
\name{font_ct}
\alias{font_ct}
\alias{font_ct_light}
\title{Crimson Text font name R variable aliases}
\format{length 1 character vector}
\usage{
font_ct

font_ct_light
}
\description{
\code{font_ct} == "\code{Crimson Text}"

\code{font_fc_light} == "\code{Crimson Text Light}"
}
\note{
\code{font_ct_light} (a.k.a. "\code{Crimson Text Light}") is not available on
Windows and will throw a warning if used in plots.
}
\keyword{datasets}
