% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humicons.R
\docType{data}
\name{font_hum}
\alias{font_hum}
\title{Humanitarian Icons font name R variable aliases}
\format{length 1 character vector}
\usage{
font_hum
}
\description{
\code{font_hum} == "\code{Humanitarian Icons}"
}
\keyword{datasets}
