% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ochathemes-package.R
\docType{package}
\name{ochathemes-package}
\alias{ochathemes-package}
\alias{ochathemes}
\title{Additional Themes and Theme Components for 'ggplot2' based on OCHA graphic styles}
\description{
A compilation of extra themes and theme components for 'ggplot2'
The core theme: `theme_ocha`
}
\author{
\email{mail@ahmadoudicko.com}
}
\keyword{package}
