% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roboto-slab.R
\docType{data}
\name{font_rsl}
\alias{font_rsl}
\alias{font_rsl_light}
\title{Roboto Slab font name R variable aliases}
\format{length 1 character vector}
\usage{
font_rsl

font_rsl_light
}
\description{
\code{font_rsl} == "\code{Roboto Slab}"

\code{font_fc_light} == "\code{Roboto Slab Light}"
}
\note{
\code{font_rsl_light} (a.k.a. "\code{Roboto Slab Light}") is not available on
Windows and will throw a warning if used in plots.
}
\keyword{datasets}
