% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_ocha}
\alias{scale_fill_ocha}
\title{#' Setup fill palette for ggplot2}
\usage{
scale_fill_ocha(..., palette = "blue", discrete = TRUE, alpha = 1,
  reverse = TRUE)
}
\arguments{
\item{...}{additional arguments to pass to scale_color_gradientn}

\item{palette}{Choose from 'ocha_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{pass through parameter to \code{viridis}}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
#' Setup fill palette for ggplot2
}
